[수정] writePage Styled 내부 정의
