+++ src/pages/WritePage.tsx
@@ -3,6 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { combinedState } from "constant/type";
 import { idAction, dateAction, resetState } from "redux/reducers/newPost";
 import { addPost } from "fb/API";
+import styled from "styled-components";
 import StyledWriteHeader from "containers/WriteHeader/WriteHeader.styled";
 import StyledTextEditor from "containers/Editor/TextEditor.styled";
 import StyledButton from "components/Button/Button.styled";
@@ -50,4 +51,20 @@ const WritePage = ({ className }: writePageProps) => {
   );
 };
 
-export default WritePage;
+const StyledWritePage = styled(WritePage)`
+  position: relative;
+  background-color: #fff;
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+  margin: 0 auto;
+
+  & > button {
+    position: fixed;
+    bottom: 10px;
+    border-radius: 15px;
+    border: 1px solid #ccc;
+  }
+`;
+
+export default StyledWritePage;
